requirements/stack.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/stack.txt requirements/stack.in
#
attrs==19.3.0             # via cattrs, jsii
aws-cdk.aws-events==1.57.0  # via aws-cdk.aws-s3
aws-cdk.aws-iam==1.57.0   # via aws-cdk.aws-events, aws-cdk.aws-kms, aws-cdk.aws-s3
aws-cdk.aws-kms==1.57.0   # via aws-cdk.aws-s3
aws-cdk.aws-s3==1.57.0    # via -r requirements/stack.in
aws-cdk.cloud-assembly-schema==1.57.0  # via aws-cdk.core, aws-cdk.cx-api
aws-cdk.core==1.57.0      # via -r requirements/stack.in, aws-cdk.aws-events, aws-cdk.aws-iam, aws-cdk.aws-kms, aws-cdk.aws-s3
aws-cdk.cx-api==1.57.0    # via aws-cdk.core
aws-cdk.region-info==1.57.0  # via aws-cdk.aws-iam
cattrs==1.0.0             # via jsii
constructs==3.0.4         # via aws-cdk.aws-events, aws-cdk.aws-iam, aws-cdk.aws-kms, aws-cdk.aws-s3, aws-cdk.core
jsii==1.10.0              # via aws-cdk.aws-events, aws-cdk.aws-iam, aws-cdk.aws-kms, aws-cdk.aws-s3, aws-cdk.cloud-assembly-schema, aws-cdk.core, aws-cdk.cx-api, aws-cdk.region-info, constructs
publication==0.0.3        # via aws-cdk.aws-events, aws-cdk.aws-iam, aws-cdk.aws-kms, aws-cdk.aws-s3, aws-cdk.cloud-assembly-schema, aws-cdk.core, aws-cdk.cx-api, aws-cdk.region-info, constructs
python-dateutil==2.8.1    # via jsii
six==1.15.0               # via python-dateutil
typing-extensions==3.7.4.2  # via jsii
